Include `proguard-base` as an API dependency of `proguard-gradle`
The public type of `ProGuardTask` exports a number of types defined in
the `proguard-base` module (`Configuration`, `ParseException`, etc). These types
are required when compiling against `ProGuardTask`, and can result in compilation
error when missing.

The correct way to model this in Gradle is to include `proguard-base` as an
API dependency of `proguard-gradle`.