Remove static cache for generated names in SimpleNameFactory (#149)
* Remove static cache for generated names

The `SimpleNameFactory` was using a static List in an attempt to improve performance,
by avoiding the need to recreate the String name multiple times. The implementation
was not thread-safe, and could result in invalid names being produced when Proguard
was being run concurrently within the same classloader.

When 2 threads entered `getName()` simultaneously, they could both construct the same
name value for the same index, and both would then add these names to the static Map.
This resulted in a name List like `['a', 'a', 'c', ...].

This bug would manifest itself with a `Duplicate Jar Entry` exception, due to the fact
that `SimpleNameFactory` would return the same name value on subsequent calls.

This commit merely removes the optimization. If name generation is demonstated to be
a performance bottleneck, and safer optimization should be implemented, ensuring
correct synchronization on shared resources.