Refactor KotlinValueParameterNameShrinker
This refactors `KotlinValueParameterNameShrinker` to remove the need for
the strange `onNewFunctionStart()` method. This will allow for removal
of `onNewFunctionStart()` since it's the only place that it's used in
DexGuard and ProGuard. `onNewFunctionStart()` does not fit into the rest
of the visitor API and the use-case is better handled as in this
refactor. Additionally the name `onNewFunctionStart()` does not make
sense since it could be a function, constructor or a property.