Mark context receiver parameters as used
To avoid the context receiver list becoming out of sync due to parameter shrinking, we now mark parameters at the same indices as context receivers as used.

This is a temporary solution and should be replaced with a solution that properly takes into account context receivers during optimization, so that e.g. unused context receivers can be removed.

~This diff adds support to the `MethodDescriptorShrinker` for shrinking the list of context receivers of a method.~

~There are some cases we cannot handle:~

~* properties where there is both a getter + setter~
~* functions with default parameters (in JVM land, a `$default` companion method)~
~* where the referencedMethod is null~

~Since we optimize each method independently, we can't be sure that the parameter lists are in sync when there are multiple JVM methods attached to a Kotlin entity like when there is a getter and setter together.~